+++ crates/ltx_cli/src/commands/mod.rs
@@ -1,11 +1,9 @@
-//! Commands is a wrapper around the CLI commands.
 pub mod check;
 pub mod clean;
 pub mod code;
 pub mod new;
 
-// re-exports
 pub use check::CheckArgs;
-pub use clean::clean_build;
-pub use code::run_code;
+pub use clean::CleanArgs;
+pub use code::CodeArgs;
 pub use new::NewArgs;