+++ CHANGELOG.md
@@ -11,11 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - added the following commands [`new`, 'clean', `code`, 'check']
+- `AppContext` and `CliCommand` to the cli context
+- `OutputFormat` Output format for status messages.
+- `manifest-path`  Path to a manifest file (`ltx.toml`) for project-level configuration.
 
 ### Changed
-- `clean` command now has cargo like functionality : `Removed 1 files, 0 B total (0.00s)`
+- `clean` command now has verbose output -v flag or for more verbose output -vv
 - the engine default is changed from `pdflatex` to `tectonic`
 - changed the out directory from `build/` to `target/`
-
+- changing the way ltx approach the cli context by introducing `AppContext` and `CliCommand`
 
 [unreleased]: https://github.com/Abdogouhmad/ltx/compare/v0.1.0...HEAD